test: Add tests for process-svg
