+++ bin/process-svg.js
@@ -1,0 +1,61 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import Svgo from 'svgo';
+import cheerio from 'cheerio';
+import { format } from 'prettier';
+
+import DEFAULT_ATTRIBUTES from '../src/default-attributes.json';
+
+/**
+ * Process SVG string.
+ * @param {string} svg - An SVG string.
+ * @param {Promise<string>}
+ */
+function processSvg(svg) {
+  return (
+    optimize(svg)
+      .then(setAttributes)
+      .then(format)
+      // remove semicolon inserted by prettier
+      // because prettier thinks it's formatting JSX not HTML
+      .then(svg => svg.replace(/;/g, ''))
+  );
+}
+
+/**
+ * Optimize SVG with `svgo`.
+ * @param {string} svg - An SVG string.
+ * @returns {Promise<string>}
+ */
+function optimize(svg) {
+  const svgo = new Svgo({
+    plugins: [
+      { convertShapeToPath: false },
+      { mergePaths: false },
+      { removeAttrs: { attrs: '(fill|stroke.*)' } },
+      { removeTitle: true },
+    ],
+  });
+
+  return new Promise((resolve, reject) => {
+    svgo.optimize(svg, ({ data }) => {
+      resolve(data);
+    });
+  });
+}
+
+/**
+ * Set default attibutes on SVG.
+ * @param {string} svg - An SVG string.
+ * @returns {string}
+ */
+function setAttributes(svg) {
+  const $ = cheerio.load(svg);
+
+  Object.keys(DEFAULT_ATTRIBUTES).forEach(key =>
+    $('svg').attr(key, DEFAULT_ATTRIBUTES[key]),
+  );
+
+  return $('body').html();
+}
+
+export default processSvg;
