--- bin/process-svgs-driver.js
@@ -1,6 +1,0 @@
-import fs from 'fs';
-import path from 'path';
-
-import processSvgs from './process-svgs';
-
-processSvgs(fs, path.resolve(__dirname, '../icons'));
