+++ package.json
@@ -7,6 +7,7 @@
     "dist"
   ],
   "scripts": {
+    "test": "jest",
     "all": "make",
     "lint": "make lint",
     "build": "make build",
@@ -27,6 +28,7 @@
     "eslint-config-airbnb-base": "^11.2.0",
     "eslint-plugin-import": "^2.5.0",
     "husky": "^0.13.4",
+    "jest": "^21.2.1",
     "parse5": "^3.0.2",
     "prettier": "^1.8.2",
     "rsvp": "^3.6.0",
